Перенос "валидации" в модуль - подчистил код
